stream_processing/requirements.txt
apache-flink==1.17.1
pyyaml==6.0.1
kafka-python==2.0.2
psycopg2-binary==2.9.7